moved drag and drop func to new file
