+++ project/index.html
@@ -5,7 +5,7 @@
   <meta name="viewport">
   <title>Fixed Grid</title>
   <link rel="stylesheet" href="sources/styles/style.css">
-  <script src="sources/scripts/main.js"></script>
+  <script src="sources/scripts/main.js" type="module"></script>
 </head>
 <body>
   <div id="grid-container">
